resolve rebase conflict with doc type enablement
