+++ app/presenters/idv/how_to_verify_presenter.rb
@@ -4,12 +4,13 @@ class Idv::HowToVerifyPresenter
   include ActionView::Helpers::TagHelper
   include ActionView::Helpers::TranslationHelper
 
-  attr_reader :selfie_required, :passport_cards_supported, :mdl_enabled
+  attr_reader :selfie_required, :passport_cards_supported, :mdl_enabled, :clear_enabled
 
-  def initialize(selfie_check_required:, passport_cards_supported: false, mdl_enabled: false)
+  def initialize(selfie_check_required:, passport_cards_supported: false, mdl_enabled: false, clear_enabled: false)
     @selfie_required = selfie_check_required
     @passport_cards_supported = passport_cards_supported
     @mdl_enabled = mdl_enabled
+    @clear_enabled = clear_enabled
   end
 
   def how_to_verify_info
