+++ app/views/idv/hybrid_handoff/show.html.erb
@@ -115,7 +115,7 @@
     </div>
   </div>
 <% end %>
-<% if true # @clear_enabled %>
+<% if @presenter.clear_enabled %>
   <hr class="margin-top-4" />
   <div class="grid-row grid-gap grid-gap-2">
     <div class="grid-col-12 tablet:grid-col-auto">
