+++ lib/idp/constants.rb
@@ -23,6 +23,7 @@ module Idp
       STATE_ID_MOCK = 'StateIdMock'
       AAMVA_CHECK_SKIPPED = 'AamvaCheckSkipped'
       SOURCE_CHECK = [AAMVA, AAMVA_UNSUPPORTED_JURISDICTION, STATE_ID_MOCK].freeze
+      CLEAR = 'clear'
     end
 
     module DocumentTypes
